Buildings super update :city_sunrise:
also literally everything else changed
+added a few TODO stuff for you to look at if you want to
+planet type specific buildings
+building types enum (makes previous work)
+different texture for space building slots, differentiates them at all now
+BuildingContainers now store isSpace
>Ship selecting is now only if the ship tab is selected
>correctly calculates max cap for buildings
>new building types take in BuildingType instead of path (why wasn't it using getImg?)
>building images are now in Images, including non added ones (acts as a reminder)
>should stop spamming .reccomenders changes from my computer in the future
*fixed lots of incorrect for loops that caused index errors
*does anyone even read these things I put a good ammount of time into?

+++ core/src/com/ue/ps/ui/SidePanel.java
@@ -189,18 +189,32 @@ public class SidePanel extends BaseActor {
 
 		//BUILDINGS TAB
 		// setup building boxes
-		for (int i = 0; i < this.planet.landBuildings.length; i++) {
-			BuildingContainer bc = new BuildingContainer(i, 10, Tab.maxHeight - (i) * 50); //from bottom left
-			//BuildingContainer bc = new BuildingContainer(i, 10, -10 - (i+1) * 50); //from top left
-			bc.planet = p;
-			this.buildingContainers.add(bc);
-			tabBuildings.addActor(bc);
-			tabBuildings.internalHeight += 50; //add the height of this object (and buffer beneath) to the net height of the tab
-
-			if (this.planet.landBuildings[i] != null) {
-				bc.setBuilding(this.planet.landBuildings[i]);
-			} else {
-				bc.setBuilding(null);
+		for (int i = 0; i < this.planet.landBuildings.length + this.planet.spaceBuildings.length; i++) {
+			if (i < this.planet.landBuildings.length) {
+				BuildingContainer bc = new BuildingContainer(i, 10, Tab.maxHeight - (i) * 50, p, false); //from bottom left
+				//BuildingContainer bc = new BuildingContainer(i, 10, -10 - (i+1) * 50); //from top left
+				this.buildingContainers.add(bc);
+				tabBuildings.addActor(bc);
+				tabBuildings.internalHeight += 50; //add the height of this object (and buffer beneath) to the net height of the tab
+	
+				if (this.planet.landBuildings[i] != null) {
+					bc.setBuilding(this.planet.landBuildings[i]);
+				} else {
+					bc.setBuilding(null);
+				}
+			} else { //Space buildings
+				BuildingContainer bc = new BuildingContainer(i, 10, Tab.maxHeight - (i) * 50, p, true); //from bottom left
+				//BuildingContainer bc = new BuildingContainer(i, 10, -10 - (i+1) * 50); //from top left
+				this.buildingContainers.add(bc);
+				tabBuildings.addActor(bc);
+				tabBuildings.internalHeight += 50; //add the height of this object (and buffer beneath) to the net height of the tab
+	
+				if (this.planet.spaceBuildings[i-this.planet.landBuildings.length] != null) {
+					bc.setBuilding(this.planet.spaceBuildings[i-this.planet.landBuildings.length]);
+				} else {
+					bc.setBuilding(null);
+				}
+				
 			}
 		}
 		
@@ -232,15 +246,12 @@ public class SidePanel extends BaseActor {
 
 		this.tabBuildings.update(uiMouseBlot.center);
 		this.tabShips.update(uiMouseBlot.center);
-		
-		// update containers
-		//for (BuildingContainer bc : this.buildingContainers) {
-		//	bc.update(uiMouseBlot.center);
-		//}
 
 		// check for clicking on increment/deincrement priority/capacity and
 		// increment/deincrement them
-		if (Gdx.input.justTouched()) {
+		
+		//TODO put in an if for the tab they will be in
+		/*if (Gdx.input.justTouched()) {
 			Rectangle mouse = uiMouseBlot.getBoundingRectangle(); // make this
 																	// global
 																	// and the
@@ -259,51 +270,52 @@ public class SidePanel extends BaseActor {
 			}
 
 			// check for destroying building
-			/*
 			 * if (this.destroyBuildingBox.getBoundingRectangle().contains(
 			 * uiMouseBlot.center) && Gdx.input.justTouched()) { //destroy
 			 * building this.planet.destroyBuilding(selectedBuildingSlot);
 			 * //update building boxes
 			 * buildingContainers.get(selectedBuildingSlot).setBuilding(null);
 			 * hideDestroy(); }
-			 */
-		}
+			 
+		}*/
 
 		// update selectedships
-
-		for (int i = 0; i < shipContainers.size(); i++) {
-			shipContainers.get(i).update(uiMouseBlot.center);
-			if (shipContainers.get(i).done) {
-				if (shipContainers.get(i).isSelected()) {
-					if (!selectedShips.contains(shipContainers.get(i).getShip())) {
-						selectedShips.add(shipContainers.get(i).getShip());
+		//TODO force deselect when tabbed off?
+		if (tabShips.selected){
+			for (int i = 0; i < shipContainers.size(); i++) {
+				shipContainers.get(i).update(uiMouseBlot.center);
+				if (shipContainers.get(i).done) {
+					if (shipContainers.get(i).isSelected()) {
+						if (!selectedShips.contains(shipContainers.get(i).getShip())) {
+							selectedShips.add(shipContainers.get(i).getShip());
+						}
+	
+					} else {
+						selectedShips.remove(shipContainers.get(i).getShip());
 					}
-
-				} else {
-					selectedShips.remove(shipContainers.get(i).getShip());
-				}
-				ShipPointer deleteThisPointer = null;
-				if (shipContainers.get(i).isDestinationUnset) {
-					for (ShipPointer sp : this.planet.pointers) {
-						sp.ships.remove(shipContainers.get(i).getShip());
-
-						if (sp.ships.isEmpty()) {
-							sp.delete();
-							deleteThisPointer = sp;
+					ShipPointer deleteThisPointer = null;
+					if (shipContainers.get(i).isDestinationUnset) {
+						for (ShipPointer sp : this.planet.pointers) {
+							sp.ships.remove(shipContainers.get(i).getShip());
+	
+							if (sp.ships.isEmpty()) {
+								sp.delete();
+								deleteThisPointer = sp;
+							}
 						}
 					}
+					this.planet.pointers.remove(deleteThisPointer);
+					if (shipContainers.get(i).getShip() != null && shipContainers.get(i).getShip().health <= 0) {
+						shipContainers.get(i).remove();
+						shipContainers.remove(i);
+						//i--;
+						
+					}
 				}
-				this.planet.pointers.remove(deleteThisPointer);
-				if (shipContainers.get(i).getShip() != null && shipContainers.get(i).getShip().health <= 0) {
-					shipContainers.get(i).remove();
-					shipContainers.remove(i);
-					//i--;
-					
-				}
+				
 			}
-			
 		}
-
+			
 		// update text fields
 		this.planetCap.setText(Integer.toString(this.planet.resourceCapacity));
 		this.planetPrioirity.setText(Integer.toString(this.planet.priority));
